Move Prometheus logging config into `prometheus` pkg
